requirements.txt
tabula-py>=2.3.0
pdfplumber>=0.7.0
pdfminer.six>=20200517
pandas>=1.3.0
Flask>=2.0.0
Markdown>=3.3.0
tabulate>=0.8.0
docling>=0.1.0
camelot-py>=0.11.0
opencv-python>=4.5.0
pypdfium2>=4.0.0
transformers>=4.30.0
torch>=2.0.0
pdf2image>=1.16.0
Pillow>=9.0.0
timm>=1.0.0
pytesseract>=0.3.0